refactor(SEP e2e tests): Separate huge SEPs tests in two files for faster run in parallel
