+++ src/components/SEP/MeetingComponents/ProposalViewModal/SEPMeetingProposalViewModal.tsx
@@ -101,6 +101,7 @@ const SEPMeetingProposalViewModal: React.FC<SEPMeetingProposalViewModalProps> = 
               color="inherit"
               onClick={handleClose}
               aria-label="close"
+              data-cy="close-modal"
             >
               <CloseIcon />
             </IconButton>
